Updated short URL list params so that it requests dateCreated DESC ordering by default
